Move SnyggValueIcon to own source file
